divisin-server/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
flask-marshmallow==1.1.0
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
marshmallow==3.20.2
packaging==23.2
pillow==10.2.0
PyMySQL==1.1.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
